Add jwt wrapper function
