+++ project/server/auth/helpers.py
@@ -1,6 +1,11 @@
-import re
+import re, jwt
+from db import DatabaseConnection
+from project.server import app
+from functools import wraps
+from flask import request, make_response, jsonify
 """ This script will contain all helper functions for authentication """
 
+db_name = DatabaseConnection()
 
 def validate_registration_input(data):
     try:
@@ -33,3 +38,47 @@ def validate_login_input(data):
                 raise TypeError("Password should be a string or an integer")
     except (TypeError, ValueError) as e:
         return str(e)
+
+def decode_auth_token(auth_token):
+    """
+    Decodes the auth token
+    :param auth_token:
+    :return: integer|string
+    """
+    try:
+        payload = jwt.decode(auth_token, app.config.get('SECRET_KEY'))
+        is_blacklisted = db_name.check_blacklist(auth_token)
+        if is_blacklisted:
+            return 'Token blacklisted. Please log in again.'
+        else:
+            return payload['sub']
+    except jwt.ExpiredSignatureError:
+        return 'Signature expired. Please log in again.'
+    except jwt.InvalidTokenError:
+        return 'Invalid token. Please log in again.'
+
+def token_required(func):
+    @wraps(func)
+    def decorated_function(*args, **kwargs):
+        auth_header = request.headers.get('Authorization')
+        if auth_header:
+            auth_token = auth_header.split(" ")[1]
+        else:
+            auth_token = ''
+        if auth_token:
+            resp = decode_auth_token(auth_token)
+            if not isinstance(resp, str):
+                return func(*args, **kwargs)
+            else:
+                responseObject = {
+                    'status': 'fail',
+                    'message': resp
+                }
+                return make_response(jsonify(responseObject)), 401
+        else:
+            responseObject = {
+                'status': 'fail',
+                'message': 'Provide a valid auth token.'
+            }
+            return make_response(jsonify(responseObject)), 403
+    return decorated_function
